refactor(providers): move shared local-option split out of openai util
openai/types.ts already imports from openai/util.ts, so importing the
option types back into util.ts closed a cycle that madge rejects.

Give the shared allowlist its own module instead: localOptions.ts ->
types.ts -> util.ts is acyclic, and util.ts drops an import it only
needed for the moved block.
